EdgeJSONStudio/requirements.txt
PySide6==6.6.1
ijson==3.2.3
json5==0.9.14
jsonschema==4.19.0
jsonpath-ng==1.5.3
jmespath==1.0.1
PyYAML==6.0.1
ruamel.yaml==0.18.6
tomli==2.0.1
tomli-w==1.0.0
